Move everything from Packager to Bootstrap
