+++ main/src/ca/uwaterloo/flix/api/Bootstrap.scala
@@ -15,16 +15,116 @@
  */
 package ca.uwaterloo.flix.api
 
+import ca.uwaterloo.flix.runtime.CompilationResult
+import ca.uwaterloo.flix.tools.{Benchmarker, Tester}
 import ca.uwaterloo.flix.tools.pkg.{FlixPackageManager, ManifestParser, MavenPackageManager}
-import ca.uwaterloo.flix.util.Result
+import ca.uwaterloo.flix.util.Formatter.AnsiTerminalFormatter
+import ca.uwaterloo.flix.util.{Options, Result, Validation}
 import ca.uwaterloo.flix.util.Result.{Err, Ok, ToOk}
+import ca.uwaterloo.flix.util.Validation.{ToFailure, ToSuccess}
 
-import java.io.PrintStream
+import java.io.{PrintStream, PrintWriter}
 import java.nio.file.attribute.BasicFileAttributes
-import java.nio.file.{FileVisitResult, Files, Path, SimpleFileVisitor}
+import java.nio.file.{FileVisitResult, Files, Path, SimpleFileVisitor, StandardOpenOption}
+import java.util.{Calendar, GregorianCalendar}
+import java.util.zip.{ZipEntry, ZipOutputStream}
 import scala.collection.mutable
+import scala.util.{Failure, Success, Using}
 
 object Bootstrap {
+
+  /**
+    * Initializes a new flix project at the given path `p`.
+    *
+    * The project must not already exist.
+    */
+  def init(p: Path, o: Options): Result[Unit, Int] = {
+    //
+    // Check that the current working directory is usable.
+    //
+    if (!Files.isDirectory(p) || !Files.isReadable(p) || !Files.isWritable(p)) {
+      throw new RuntimeException(s"The directory: '$p' is not accessible. Aborting.")
+    }
+
+    //
+    // Compute the name of the package based on the directory name.
+    //
+    val packageName = getPackageName(p)
+
+    //
+    // Compute all the directories and files we intend to create.
+    //
+    val buildDirectory = getBuildDirectory(p)
+    val libraryDirectory = getLibraryDirectory(p)
+    val sourceDirectory = getSourceDirectory(p)
+    val testDirectory = getTestDirectory(p)
+
+    val manifestFile = getManifestFile(p)
+    val licenseFile = getLicenseFile(p)
+    val readmeFile = getReadmeFile(p)
+    val mainSourceFile = getMainSourceFile(p)
+    val mainTestFile = getMainTestFile(p)
+
+    //
+    // Check that the project directories and files do not already exist.
+    //
+    val dirPaths = List(
+      buildDirectory, libraryDirectory, sourceDirectory, testDirectory
+    )
+
+    // blocked directories are those that already exist and are not directories
+    val blockedDirs = dirPaths.filter(p => Files.exists(p) && !Files.isDirectory(p))
+
+    if (blockedDirs.nonEmpty) {
+      throw new RuntimeException(s"The following files already exist and are not directories: ${blockedDirs.mkString(", ")}. Aborting.")
+    }
+
+    //
+    // Create the project directories and files.
+    //
+    newDirectoryIfAbsent(buildDirectory)
+    newDirectoryIfAbsent(libraryDirectory)
+    newDirectoryIfAbsent(sourceDirectory)
+    newDirectoryIfAbsent(testDirectory)
+
+    newFileIfAbsent(manifestFile) {
+      s"""[package]
+         |name    = "$packageName"
+         |description = "test"
+         |version = "0.1.0"
+         |flix    = "${Version.CurrentVersion}"
+         |authors = ["Tester"]
+         |""".stripMargin
+    }
+
+    newFileIfAbsent(licenseFile) {
+      """Enter license information here.
+        |""".stripMargin
+    }
+
+    newFileIfAbsent(readmeFile) {
+      s"""# $packageName
+         |
+         |Enter some useful information.
+         |
+         |""".stripMargin
+    }
+
+    newFileIfAbsent(mainSourceFile) {
+      """// The main entry point.
+        |def main(): Unit \ IO =
+        |    println("Hello World!")
+        |""".stripMargin
+    }
+
+    newFileIfAbsent(mainTestFile) {
+      """@test
+        |def test01(): Bool = 1 + 1 == 2
+        |""".stripMargin
+    }
+    ().toOk
+  }
+
 
   /**
     * Returns the path to the library directory relative to the given path `p`.
@@ -34,22 +134,137 @@ object Bootstrap {
   /**
     * Returns the path to the source directory relative to the given path `p`.
     */
-  def getSourceDirectory(p: Path): Path = p.resolve("./src/").normalize()
+  private def getSourceDirectory(p: Path): Path = p.resolve("./src/").normalize()
 
   /**
     * Returns the path to the test directory relative to the given path `p`.
     */
-  def getTestDirectory(p: Path): Path = p.resolve("./test/").normalize()
+  private def getTestDirectory(p: Path): Path = p.resolve("./test/").normalize()
 
   /**
     * Returns the path to the build directory relative to the given path `p`.
     */
-  def getBuildDirectory(p: Path): Path = p.resolve("./build/").normalize()
+  private def getBuildDirectory(p: Path): Path = p.resolve("./build/").normalize()
+
+  /**
+    * Returns the path to the LICENSE file relative to the given path `p`.
+    */
+  private def getLicenseFile(p: Path): Path = p.resolve("./LICENSE.md").normalize()
+
+  /**
+    * Returns the path to the README file relative to the given path `p`.
+    */
+  private def getReadmeFile(p: Path): Path = p.resolve("./README.md").normalize()
+
+  /**
+    * Returns the path to the main source file relative to the given path `p`.
+    */
+  private def getMainSourceFile(p: Path): Path = getSourceDirectory(p).resolve("./Main.flix").normalize()
+
+  /**
+    * Returns the path to the main test file relative to the given path `p`.
+    */
+  private def getMainTestFile(p: Path): Path = getTestDirectory(p).resolve("./TestMain.flix").normalize()
 
   /**
     * Returns the path to the Manifest file relative to the given path `p`.
     */
-  def getManifestFile(p: Path): Path = p.resolve("./flix.toml").normalize()
+  private def getManifestFile(p: Path): Path = p.resolve("./flix.toml").normalize()
+
+  /**
+    * Returns the path to the jar file based on the given path `p`.
+    */
+  private def getJarFile(p: Path): Path = p.resolve(getPackageName(p) + ".jar").normalize()
+
+  /**
+    * Returns the package name based on the given path `p`.
+    */
+  private def getPackageName(p: Path): String = p.toAbsolutePath.normalize().getFileName.toString
+
+  /**
+    * Returns the path to the pkg file based on the given path `p`.
+    */
+  private def getPkgFile(p: Path): Path = p.resolve(getPackageName(p) + ".fpkg").normalize()
+
+  /**
+    * Returns `true` if the given path `p` is a jar-file.
+    */
+  private def isJarFile(p: Path): Boolean = p.getFileName.toString.endsWith(".jar") && isZipArchive(p)
+
+  /**
+    * Returns `true` if the given path `p` is a fpkg-file.
+    */
+  def isPkgFile(p: Path): Boolean = p.getFileName.toString.endsWith(".fpkg") && isZipArchive(p)
+
+  /**
+    * Creates a new directory at the given path `p`.
+    * The path must not already contain a non-directory.
+    */
+  private def newDirectoryIfAbsent(p: Path): Unit = {
+    Files.createDirectories(p)
+  }
+
+  /**
+    * Creates a new text file at the given path `p` with the given content `s`,
+    * if the file does not already exist.
+    */
+  private def newFileIfAbsent(p: Path)(s: String): Unit = {
+    Files.writeString(p, s, StandardOpenOption.CREATE)
+  }
+
+  /**
+    * To support DOS time, Java 8+ treats dates before the 1980 January in special way.
+    * Here we use 2014-06-27 (the date of the first commit to Flix) to avoid the complexity introduced by this hack.
+    *
+    * @see <a href="https://bugs.openjdk.java.net/browse/JDK-4759491">JDK-4759491 that introduced the hack around 1980 January from Java 8+</a>
+    * @see <a href="https://bugs.openjdk.java.net/browse/JDK-6303183">JDK-6303183 that explains why the second should be even to create ZIP files in platform-independent way</a>
+    * @see <a href="https://github.com/gradle/gradle/blob/445deb9aa988e506120b7918bf91acb421e429ba/subprojects/core/src/main/java/org/gradle/api/internal/file/archive/ZipCopyAction.java#L42-L57">A similar case from Gradle</a>
+    */
+  private val ENOUGH_OLD_CONSTANT_TIME: Long = new GregorianCalendar(2014, Calendar.JUNE, 27, 0, 0, 0).getTimeInMillis
+
+  /**
+    * Adds an entry to the given zip file.
+    */
+  private def addToZip(zip: ZipOutputStream, name: String, p: Path): Unit = {
+    addToZip(zip, name, Files.readAllBytes(p))
+  }
+
+  /**
+    * Adds an entry to the given zip file.
+    */
+  private def addToZip(zip: ZipOutputStream, name: String, d: Array[Byte]): Unit = {
+    val entry = new ZipEntry(name)
+    entry.setTime(ENOUGH_OLD_CONSTANT_TIME)
+    zip.putNextEntry(entry)
+    zip.write(d)
+    zip.closeEntry()
+  }
+
+  /**
+    * Returns `true` if the given path `p` is a zip-archive.
+    */
+  private def isZipArchive(p: Path): Boolean = {
+    if (Files.exists(p) && Files.isReadable(p) && Files.isRegularFile(p)) {
+      // Read the first four bytes of the file.
+      return Using(Files.newInputStream(p)) { is =>
+        val b1 = is.read()
+        val b2 = is.read()
+        val b3 = is.read()
+        val b4 = is.read()
+        // Check if the four first bytes match 0x50, 0x4b, 0x03, 0x04
+        return b1 == 0x50 && b2 == 0x4b && b3 == 0x03 && b4 == 0x04
+      }.get
+    }
+    false
+  }
+
+  /**
+    * @param root the root directory to compute a relative path from the given path
+    * @param path the path to be converted to a relative path based on the given root directory
+    * @return relative file name separated by slashes, like `path/to/file.ext`
+    */
+  private def convertPathToRelativeFileName(root: Path, path: Path): String =
+    root.relativize(path).toString.replace('\\', '/')
 
   /**
     * Returns all files in the given path `p` ending with .`ext`.
@@ -60,7 +275,7 @@ object Bootstrap {
   /**
     * Returns all files in the given path `p`.
     */
-  def getAllFiles(p: Path): List[Path] = {
+  private def getAllFiles(p: Path): List[Path] = {
     if (Files.isReadable(p) && Files.isDirectory(p)) {
       val visitor = new FileVisitor
       Files.walkFileTree(p, visitor)
@@ -104,19 +319,19 @@ object Bootstrap {
     // Determine the mode: If `path/flix.toml` exists then "project" mode else "folder mode".
     //
     val bootstrap = new Bootstrap(path)
-    val tomlPath = Bootstrap.getManifestFile(path)
+    val tomlPath = getManifestFile(path)
     if (Files.exists(tomlPath)) {
       out.println("Found `flix.toml'. Checking dependencies...")
-      bootstrap.projectMode(path).map(_ => bootstrap)
+      bootstrap.projectMode().map(_ => bootstrap)
     } else {
       out.println("No `flix.toml'. Will load source files from `*.flix`, `src/**`, and `test/**`.")
-      bootstrap.folderMode(path).map(_ => bootstrap)
+      bootstrap.folderMode().map(_ => bootstrap)
     }
   }
 
 }
 
-class Bootstrap(val path: Path) {
+class Bootstrap(val projectPath: Path) {
 
   // Timestamps at the point the sources were loaded
   private var timestamps: Map[Path, Long] = Map.empty
@@ -131,29 +346,29 @@ class Bootstrap(val path: Path) {
     * Then makes a list of all flix source files, flix packages
     * and .jar files that this project uses.
     */
-  private def projectMode(path: Path)(implicit out: PrintStream): Result[Unit, BootstrapError] = {
+  private def projectMode()(implicit out: PrintStream): Result[Unit, BootstrapError] = {
     // 1. Read, parse, and validate flix.toml.
-    val tomlPath = Bootstrap.getManifestFile(path)
+    val tomlPath = Bootstrap.getManifestFile(projectPath)
     val manifest = ManifestParser.parse(tomlPath) match {
       case Ok(m) => m
       case Err(e) => return Err(BootstrapError.ManifestParseError(e))
     }
 
     // 2. Check each dependency is available or download it.
-    FlixPackageManager.installAll(manifest, path) match {
+    FlixPackageManager.installAll(manifest, projectPath) match {
       case Ok(l) => flixPackagePaths = l
       case Err(e) => return Err(BootstrapError.FlixPackageError(e))
     }
-    MavenPackageManager.installAll(manifest, path) match {
+    MavenPackageManager.installAll(manifest, projectPath) match {
       case Ok(l) => mavenPackagePaths = l
       case Err(e) => return Err(BootstrapError.MavenPackageError(e))
     }
     out.println("Dependency resolution completed.")
 
     // 3. Add *.flix, src/**.flix and test/**.flix
-    val filesHere = Bootstrap.getAllFlixFilesHere(path)
-    val filesSrc = Bootstrap.getAllFilesWithExt(Bootstrap.getSourceDirectory(path), "flix")
-    val filesTest = Bootstrap.getAllFilesWithExt(Bootstrap.getTestDirectory(path), "flix")
+    val filesHere = Bootstrap.getAllFlixFilesHere(projectPath)
+    val filesSrc = Bootstrap.getAllFilesWithExt(Bootstrap.getSourceDirectory(projectPath), "flix")
+    val filesTest = Bootstrap.getAllFilesWithExt(Bootstrap.getTestDirectory(projectPath), "flix")
     sourcePaths = filesHere ++ filesSrc ++ filesTest
 
     ().toOk
@@ -164,19 +379,19 @@ class Bootstrap(val path: Path) {
     * Then makes a list of all flix source files, flix packages
     * and .jar files that this project uses.
     */
-  private def folderMode(path: Path): Result[Unit, BootstrapError] = {
+  private def folderMode(): Result[Unit, BootstrapError] = {
     // 1. Add *.flix, src/**.flix and test/**.flix
-    val filesHere = Bootstrap.getAllFlixFilesHere(path)
-    val filesSrc = Bootstrap.getAllFilesWithExt(Bootstrap.getSourceDirectory(path), "flix")
-    val filesTest = Bootstrap.getAllFilesWithExt(Bootstrap.getTestDirectory(path), "flix")
+    val filesHere = Bootstrap.getAllFlixFilesHere(projectPath)
+    val filesSrc = Bootstrap.getAllFilesWithExt(Bootstrap.getSourceDirectory(projectPath), "flix")
+    val filesTest = Bootstrap.getAllFilesWithExt(Bootstrap.getTestDirectory(projectPath), "flix")
     sourcePaths = filesHere ++ filesSrc ++ filesTest
 
     // 2. Grab all jars in lib/
-    val jarFilesLib = Bootstrap.getAllFilesWithExt(Bootstrap.getLibraryDirectory(path), "jar")
+    val jarFilesLib = Bootstrap.getAllFilesWithExt(Bootstrap.getLibraryDirectory(projectPath), "jar")
     mavenPackagePaths = jarFilesLib
 
     // 3. Grab all flix packages in lib/
-    val flixFilesLib = Bootstrap.getAllFilesWithExt(Bootstrap.getLibraryDirectory(path), "fpkg")
+    val flixFilesLib = Bootstrap.getAllFilesWithExt(Bootstrap.getLibraryDirectory(projectPath), "fpkg")
     flixPackagePaths = flixFilesLib
 
     ().toOk
@@ -217,5 +432,182 @@ class Bootstrap(val path: Path) {
     */
   private def hasChanged(file: Path) = {
     !(timestamps contains file) || (timestamps(file) != file.toFile.lastModified())
+  }
+
+  /**
+    * Type checks the source files for the project.
+    */
+  def check(o: Options): Result[Unit, Int] = {
+    // Configure a new Flix object.
+    implicit val flix: Flix = new Flix()
+    flix.setOptions(o)
+
+    // Add sources and packages.
+    reconfigureFlix(flix)
+
+    flix.check() match {
+      case Validation.Success(_) => ().toOk
+      case failure =>
+        flix.mkMessages(failure.errors).foreach(println)
+        Result.Err(1)
+    }
+  }
+
+  /**
+    * Builds (compiles) the source files for the project.
+    */
+  def build(o: Options, loadClasses: Boolean = true)(implicit flix: Flix = new Flix()): Result[CompilationResult, Int] = {
+    // Configure a new Flix object.
+    val newOptions = o.copy(
+      output = Some(Bootstrap.getBuildDirectory(projectPath)),
+      loadClassFiles = loadClasses
+    )
+    flix.setOptions(newOptions)
+
+    // Add sources and packages.
+    reconfigureFlix(flix)
+
+    flix.compile() match {
+      case Validation.Success(r) => Result.Ok(r)
+      case failure =>
+        flix.mkMessages(failure.errors).foreach(println)
+        Err(1)
+    }
+  }
+
+  /**
+    * Runs the main function in flix package for the project.
+    */
+  def run(o: Options): Result[Unit, Int] = {
+    val res = for {
+      compilationResult <- build(o).toOption
+      main <- compilationResult.getMain
+    } yield {
+      main(Array.empty)
+      ().toOk[Unit, Int]
+    }
+    res.getOrElse(Err(1))
+  }
+
+  /**
+    * Runs all benchmarks in the flix package for the project.
+    */
+  def benchmark(o: Options): Result[Unit, Int] = {
+    build(o) map {
+      compilationResult =>
+        Benchmarker.benchmark(compilationResult, new PrintWriter(System.out, true))(o)
+    }
+  }
+
+  /**
+    * Runs all tests in the flix package for the project.
+    */
+  def test(o: Options): Result[Unit, Int] = {
+    implicit val flix: Flix = new Flix().setFormatter(AnsiTerminalFormatter)
+    build(o) flatMap {
+      compilationResult =>
+        Tester.run(Nil, compilationResult)
+        ().toOk
+    }
+  }
+
+  /**
+    * Builds a jar package for the project.
+    */
+  def buildJar(o: Options): Result[Unit, Int] = {
+    // The path to the jar file.
+    val jarFile = Bootstrap.getJarFile(projectPath)
+
+    // Check whether it is safe to write to the file.
+    if (Files.exists(jarFile) && !Bootstrap.isJarFile(jarFile)) {
+      throw new RuntimeException(s"The path '$jarFile' exists and is not a jar-file. Refusing to overwrite.")
+    }
+
+    // Construct a new zip file.
+    Using(new ZipOutputStream(Files.newOutputStream(jarFile))) { zip =>
+      // META-INF/MANIFEST.MF
+      val manifest =
+        """Manifest-Version: 1.0
+          |Main-Class: Main
+          |""".stripMargin
+
+      // Add manifest file.
+      Bootstrap.addToZip(zip, "META-INF/MANIFEST.MF", manifest.getBytes)
+
+      // Add all class files.
+      // Here we sort entries by relative file name to apply https://reproducible-builds.org/
+      for ((buildFile, fileNameWithSlashes) <- Bootstrap.getAllFiles(Bootstrap.getBuildDirectory(projectPath))
+        .map { path => (projectPath, Bootstrap.convertPathToRelativeFileName(Bootstrap.getBuildDirectory(projectPath), path)) }
+        .sortBy(_._2)) {
+        Bootstrap.addToZip(zip, fileNameWithSlashes, buildFile)
+      }
+    } match {
+      case Success(()) => ().toOk
+      case Failure(e) =>
+        println(e.getMessage)
+        Err(1)
+    }
+  }
+
+  /**
+    * Builds a flix package for the project.
+    */
+  def buildPkg(o: Options): Result[Unit, Int] = {
+    //
+    // We require that a Flix package has a specific structure.
+    //
+    checkProjectPath(projectPath) match {
+      case Validation.Success(_) => ()
+      case failure => throw new RuntimeException(s"Missing files or directories: ${failure.errors.mkString(", ")}")
+    }
+
+    // The path to the fpkg file.
+    val pkgFile = Bootstrap.getPkgFile(projectPath)
+
+    // Check whether it is safe to write to the file.
+    if (Files.exists(pkgFile) && !Bootstrap.isPkgFile(pkgFile)) {
+      throw new RuntimeException(s"The path '$pkgFile' exists and is not a fpkg-file. Refusing to overwrite.")
+    }
+
+    // Construct a new zip file.
+    Using(new ZipOutputStream(Files.newOutputStream(pkgFile))) { zip =>
+      // Add required resources.
+      Bootstrap.addToZip(zip, "LICENSE.md", Bootstrap.getLicenseFile(projectPath))
+      Bootstrap.addToZip(zip, "README.md", Bootstrap.getReadmeFile(projectPath))
+
+      // Add all source files.
+      // Here we sort entries by relative file name to apply https://reproducible-builds.org/
+      for ((sourceFile, fileNameWithSlashes) <- Bootstrap.getAllFiles(Bootstrap.getSourceDirectory(projectPath))
+        .map { path => (projectPath, Bootstrap.convertPathToRelativeFileName(projectPath, path)) }
+        .sortBy(_._2)) {
+        Bootstrap.addToZip(zip, fileNameWithSlashes, sourceFile)
+      }
+    } match {
+      case Success(()) => ().toOk
+      case Failure(e) =>
+        println(e.getMessage)
+        Err(1)
+    }
+  }
+
+  /**
+    * Returns a Validation containing the list of missing paths in case the path is not a project path.
+    */
+  //TODO: delete?
+  private def checkProjectPath(p: Path): Validation[Unit, Path] = {
+    val required = List(
+      Bootstrap.getSourceDirectory(p),
+      Bootstrap.getTestDirectory(p),
+      Bootstrap.getLicenseFile(p),
+      Bootstrap.getReadmeFile(p)
+    )
+    Validation.traverseX(required) {
+      case path =>
+        if (Files.exists(path)) {
+          ().toSuccess
+        } else {
+          path.toFailure
+        }
+    }
   }
 }
