+++ main/test/ca/uwaterloo/flix/tools/TestBootstrap.scala
@@ -12,7 +12,7 @@ import java.util.zip.ZipFile
 import scala.jdk.CollectionConverters.EnumerationHasAsScala
 import scala.util.Using
 
-class TestPackager extends FunSuite {
+class TestBootstrap extends FunSuite {
 
   private val ProjectPrefix: String = "flix-project-"
 
@@ -20,38 +20,38 @@ class TestPackager extends FunSuite {
 
   test("init") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
   }
 
   test("check") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.check(b, DefaultOptions)
+    b.check(DefaultOptions)
   }
 
   test("build") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.build(b, DefaultOptions)
+    b.build(DefaultOptions)
   }
 
   test("build-jar") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.build(b, DefaultOptions)
-    Packager.buildJar(b, DefaultOptions)
+    b.build(DefaultOptions)
+    b.buildJar(DefaultOptions)
 
     val packageName = p.getFileName.toString
     val jarPath = p.resolve(packageName + ".jar")
@@ -61,13 +61,13 @@ class TestPackager extends FunSuite {
 
   test("build-jar generates ZIP entries with fixed time") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.build(b, DefaultOptions)
-    Packager.buildJar(b, DefaultOptions)
+    b.build(DefaultOptions)
+    b.buildJar(DefaultOptions)
 
     val packageName = p.getFileName.toString
     val packagePath = p.resolve(packageName + ".jar")
@@ -81,7 +81,7 @@ class TestPackager extends FunSuite {
 
   test("build-jar always generates package that is byte-for-byte exactly the same") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val packageName = p.getFileName.toString
     val packagePath = p.resolve(packageName + ".jar")
 
@@ -89,12 +89,12 @@ class TestPackager extends FunSuite {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.build(b, DefaultOptions)
-    Packager.buildJar(b, DefaultOptions)
+    b.build(DefaultOptions)
+    b.buildJar(DefaultOptions)
     def hash1 = calcHash(packagePath)
 
-    Packager.build(b, DefaultOptions)
-    Packager.buildJar(b, DefaultOptions)
+    b.build(DefaultOptions)
+    b.buildJar(DefaultOptions)
     def hash2 = calcHash(packagePath)
 
     assert(
@@ -104,13 +104,13 @@ class TestPackager extends FunSuite {
 
   test("build-pkg") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
 
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.buildPkg(b, DefaultOptions)
+    b.buildPkg(DefaultOptions)
 
     val packageName = p.getFileName.toString
     val packagePath = p.resolve(packageName + ".fpkg")
@@ -120,13 +120,13 @@ class TestPackager extends FunSuite {
 
   test("build-pkg generates ZIP entries with fixed time") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
 
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.buildPkg(b, DefaultOptions)
+    b.buildPkg(DefaultOptions)
 
     val packageName = p.getFileName.toString
     val packagePath = p.resolve(packageName + ".fpkg")
@@ -140,7 +140,7 @@ class TestPackager extends FunSuite {
 
   test("build-pkg always generates package that is byte-for-byte exactly the same") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val packageName = p.getFileName.toString
     val packagePath = p.resolve(packageName + ".fpkg")
 
@@ -148,10 +148,10 @@ class TestPackager extends FunSuite {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.buildPkg(b, DefaultOptions)
+    b.buildPkg(DefaultOptions)
     def hash1 = calcHash(packagePath)
 
-    Packager.buildPkg(b, DefaultOptions)
+    b.buildPkg(DefaultOptions)
     def hash2 = calcHash(packagePath)
 
     assert(
@@ -161,32 +161,32 @@ class TestPackager extends FunSuite {
 
   test("benchmark") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.benchmark(b, DefaultOptions)
+    b.benchmark(DefaultOptions)
   }
 
   test("run") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.run(b, DefaultOptions)
+    b.run(DefaultOptions)
   }
 
   test("test") {
     val p = Files.createTempDirectory(ProjectPrefix)
-    Packager.init(p, DefaultOptions)
+    Bootstrap.init(p, DefaultOptions)
     val b = Bootstrap.bootstrap(p)(System.out) match {
       case Result.Ok(bootstrap) => bootstrap
       case Result.Err(e) => fail(e.toString)
     }
-    Packager.test(b, DefaultOptions)
+    b.test(DefaultOptions)
   }
 
   def calcHash(p: Path): String = {
