+++ main/test/ca/uwaterloo/flix/tools/ToolsSuite.scala
@@ -3,7 +3,7 @@ package ca.uwaterloo.flix.tools
 import org.scalatest.Suites
 
 class ToolsSuite extends Suites(
-  new TestPackager,
+  new TestBootstrap,
   new TestManifestParser
 ) {
   /* left empty */
